Move cause generation into InitErrorObjectNode. Moved InstallErrorCauseNode out of ConstructorBuiltins.
